rename: "directory -> document"
