+++ main.css
@@ -6,7 +6,7 @@ li {
   list-style: none;
 }
 
-div .directory-page {
+div .document-page {
   width: 300px;
   margin-right: 10px;
 }
